许哲：筛查APP相关

See merge request web/myopia-pac/myopia-pac-backend!519

+++ myopia-business/business-core/screening-flow-core/src/main/java/com/wupol/myopia/business/core/screening/flow/domain/dos/BloodPressureDataDO.java
@@ -6,6 +6,11 @@ import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
 
 import java.io.Serializable;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * @Description 血压
@@ -20,10 +25,10 @@ public class BloodPressureDataDO extends AbstractDiagnosisResult implements Seri
     /**
      * 舒张压
      */
-    private Float dbp;
+    private BigDecimal dbp;
 
     /**
      * 收缩压
      */
-    private Float sbp;
+    private BigDecimal sbp;
 }
